test3

+++ src/mx/itson/cinepolis/entidades/Cliente.java
@@ -12,7 +12,7 @@ import java.time.LocalDate;
  */
 public class Cliente {
     
-    private int id;
+    private long id;
     private String nombre;
     private String apellido;
     private boolean recibirPublicidad;
@@ -24,7 +24,7 @@ public class Cliente {
     public Cliente() {
     }
 
-    public Cliente(int id, String nombre, String apellido, boolean recibirPublicidad, String correo, LocalDate fechaCumpleanos, String ciudad, String contrasenia) {
+    public Cliente(long id, String nombre, String apellido, boolean recibirPublicidad, String correo, LocalDate fechaCumpleanos, String ciudad, String contrasenia) {
         this.id = id;
         this.nombre = nombre;
         this.apellido = apellido;
@@ -34,7 +34,6 @@ public class Cliente {
         this.ciudad = ciudad;
         this.contrasenia = contrasenia;
     }
-
 
     @Override
     public String toString() {
@@ -52,11 +51,11 @@ public class Cliente {
         return sb.toString();
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -115,5 +114,5 @@ public class Cliente {
     public void setContrasenia(String contrasenia) {
         this.contrasenia = contrasenia;
     }
-
+    
 }
